Feat : S3적용을 위한 기존 파일들 수정(#8)
1. 관리자 인덱스 페이지 CSS적용
2. S3적용(img -> imgUrl)
3. 카테고리 타입 에러 해결

+++ jumun/src/main/java/com/mihak/jumun/category/CategoryRepository.java
@@ -4,8 +4,8 @@ import com.mihak.jumun.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
-
-public interface CategoryRepository extends JpaRepository<Category,Integer> {
+/*TypeMismatchException 에러 해결*/
+public interface CategoryRepository extends JpaRepository<Category,Long> {
     Optional<Category> findByName(String name);
 
 }
